Change base type and add new numeric inputs

+++ SmartCache/sim/cache/functions.go
@@ -16,14 +16,14 @@ const (
 	FuncWeightedRequests
 )
 
-func fileWeightedAdditiveFunction(totRequests uint32, size float32, meanTicks float32, alpha float32, beta float32, gamma float32) float32 {
-	return alpha*float32(totRequests) + beta*size + gamma*meanTicks
+func fileWeightedAdditiveFunction(totRequests int64, size float64, meanTicks float64, alpha float64, beta float64, gamma float64) float64 {
+	return alpha*float64(totRequests) + beta*size + gamma*meanTicks
 }
 
-func fileWeightedMultiplicativeFunction(totRequests uint32, size float32, meanTicks float32, alpha float32, beta float32, gamma float32) float32 {
-	return float32(math.Pow(float64(totRequests), float64(alpha)) * math.Pow(float64(size), float64(beta)) * math.Pow(float64(meanTicks), float64(gamma)))
+func fileWeightedMultiplicativeFunction(totRequests int64, size float64, meanTicks float64, alpha float64, beta float64, gamma float64) float64 {
+	return float64(math.Pow(float64(totRequests), alpha) * math.Pow(float64(size), beta) * math.Pow(float64(meanTicks), gamma))
 }
 
-func fileWeightedRequest(totRequests uint32, size float32, meanTicks float32) float32 {
-	return meanTicks + (size / float32(math.Exp(float64(totRequests))))
+func fileWeightedRequest(totRequests int64, size float64, meanTicks float64) float64 {
+	return meanTicks + (size / math.Exp(float64(totRequests)))
 }
